requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Image processing (use headless for server deployment)
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
rasterio>=1.3.0
matplotlib>=3.7.0
scikit-image>=0.21.0

# PyTorch (CPU version for server)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
